Don't attempt to show massif output in quick mode.

+++ test/run.sh
@@ -125,7 +125,7 @@ else
 fi
 echo "--------------------------------------------"
 
-if [ `which valgrind` ] && [ -z "${LG_SMOKE}" ]; then
+if [ `which valgrind` ] && [ -z "${LG_SMOKE}" ] && [ -z "${LG_QUICK}" ]; then
     echo Peak stack in $(grep -h mem_stacks_B massif.out.* | sort -h -k 2 -t '=' | tail -1 | sed -e 's/mem_stacks_B=//') bytes
     echo Peak heap in $(grep -h mem_heap_B massif.out.* | sort -h -k 2 -t '=' | tail -1 | sed -e 's/mem_heap_B=//') bytes
     rm -f massif.out.*
